requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
requests==2.32.3
pypdf==5.3.1
python-docx==1.1.2
pillow==11.1.0
pytesseract==0.3.13
cryptography==46.0.3
